Python/ml-powered-applications/neel/requirements.txt
pandas==1.1.5
scikit_learn==0.24.1
bokeh==2.2.3
matplotlib==3.3.3
PyYAML==5.4.1
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
umap-learn==0.5.0
joblib==1.0.0
beautifulsoup4==4.9.3